+++ lib/cuprum/errors.rb
@@ -1,0 +1,6 @@
+require 'cuprum'
+
+module Cuprum
+  # Namespace for custom Cuprum error classes.
+  module Errors; end
+end