+++ lib/cuprum/errors/process_not_implemented_error.rb
@@ -1,14 +1,14 @@
-require 'cuprum'
+require 'cuprum/errors'
 
-module Cuprum
+module Cuprum::Errors
   # Error class for calling a Command that was not given a definition block
   # or have a #process method defined.
-  class NotImplementedError < StandardError
-    # Error message for a NotImplementedError.
+  class ProcessNotImplementedError < StandardError
+    # Error message for a ProcessNotImplementedError.
     DEFAULT_MESSAGE = 'no implementation defined for command'.freeze
 
     def initialize message = nil
       super(message || DEFAULT_MESSAGE)
-    end # constructor
-  end # class
-end # module
+    end
+  end
+end